Add an option to build without the minify plugin
This relies an inheritance <https://www.mkdocs.org/user-guide/configuration/\#configuration-inheritance\>
to define different config files for different settings.

The 'mkdocs build' and 'mkdocs serve' are unchanged (and 'make build'
is now an alias of the former); and this add a 'make build-dev' command
that runs without the minify plugin.

This saves over 1s of a ~5.5s run time on my machine.