Convert docs/3/commands/ to yaml
This should help with GH-107.

This does not change the HTML output, except for the removed
HTML comment (`<!-- This file contains a page fragment. Any changes will affect all pages that include it. -->`)

For the record, I used this script:

```
import glob
import pprint
import re
import textwrap

for filename in sorted(glob.glob("docs/3/commands/_*.md")):
    cmd_name = re.match("docs/3/commands/_cmd_(.*).md", filename).group(1)
    print(cmd_name)
    with open(filename) as fd:
        content = fd.read()
    m = re.match(r".*\n### `(?P<syntax>[^`]+)`\n+(?P<description>[^#]*)\n+(#### Example Usage\n+(?P<example>.*))?$", content, re.DOTALL | re.MULTILINE)
    pprint.pprint(m.groupdict())

    with open(f"docs/3/commands/_{cmd_name}.yml", "wt") as fd:
        fd.write(textwrap.dedent(
            f"""
            name: {cmd_name}
            syntax: >-
              {m.group("syntax")}
            description: |-
            """
        ).lstrip())

        fd.write(textwrap.indent(m.group("description"), "  "))

        if m.group("example"):
            fd.write("example: |-\n" + textwrap.indent(m.group("example"), "  "))
```

plus a manual edit to `_stats.yml` because the regexp could not handle the
extra subtitle, and added newlines at the end of files missing one.